Fix imports

+++ clinica/pipelines/t1_freesurfer_longitudinal/t1_freesurfer_template_utils.py
@@ -31,8 +31,6 @@ def init_input_node(caps_dir, participant_id, list_session_ids, output_dir):
         if e.errno != errno.EEXIST:  # EEXIST: folder already exists
             raise e
 
-    from clinica.utils.stream import cprint
-
     # Create symbolic links containing cross-sectional segmentation(s) in SUBJECTS_DIR so that recon-all can run
     for session_id in list_session_ids:
         cross_sectional_path = os.path.join(
@@ -49,7 +47,6 @@ def init_input_node(caps_dir, participant_id, list_session_ids, output_dir):
         except FileExistsError as e:
             if e.errno != errno.EEXIST:  # EEXIST: folder already exists
                 raise e
-        cprint('Creating sym link from %s to %s' % (cross_sectional_path, subjects_dir))
 
     # Prepare arguments for recon-all.
     flags = ""
@@ -81,11 +78,9 @@ def run_recon_all_base(subjects_dir,
     See official documentation (https://surfer.nmr.mgh.harvard.edu/fswiki/LongitudinalProcessing) for details.
     """
     import subprocess
-    from clinica.utils.stream import cprint
 
     recon_all_base_command = 'recon-all -base {0} {1} -sd {2} {3}'.format(
         subject_id, flags, subjects_dir, directive)
-    cprint(recon_all_base_command)
     subprocess_run_recon_all_base = subprocess.run(
         recon_all_base_command,
         shell=True,
@@ -111,7 +106,6 @@ def save_to_caps(source_dir, image_id, list_session_ids, caps_dir, overwrite_cap
         If an error occurs, it will be detected by Nipype and the next nodes (i.e. save_to_caps will not be called).
     Raise:
         FileNotFoundError: If symbolic links in `source_dir`/`image_id` folder are not removed
-        IOError: If the `source_dir`/`image_id` folder does not contain FreeSurfer segmentation.
     """
     import os
     import datetime
@@ -119,6 +113,7 @@ def save_to_caps(source_dir, image_id, list_session_ids, caps_dir, overwrite_cap
     import shutil
     from colorama import Fore
     from clinica.utils.stream import cprint
+    from clinica.utils.longitudinal import save_long_id
     from clinica.utils.ux import print_end_image
 
     participant_id = image_id.split('_')[0]
@@ -129,30 +124,40 @@ def save_to_caps(source_dir, image_id, list_session_ids, caps_dir, overwrite_cap
         'subjects',
         participant_id,
         long_id,
-        't1',
         'freesurfer_unbiased_template'
     )
 
+    # Save <long_id>_sessions.tsv to retrieve sessions used to generate <long_id>
+    sessions_tsv_path = os.path.join(
+        os.path.expanduser(caps_dir),
+        'subjects',
+        participant_id,
+        long_id
+    )
+    if not os.path.isfile(os.path.join(sessions_tsv_path, long_id + '_sessions.tsv')):
+        save_long_id(list_session_ids, sessions_tsv_path, long_id + '_sessions.tsv')
+
+    # Save FreeSurfer segmentation
     representative_file = os.path.join(image_id, 'mri', 'aparc+aseg.mgz')
     representative_source_file = os.path.join(os.path.expanduser(source_dir), image_id, representative_file)
     representative_destination_file = os.path.join(destination_dir, representative_file)
     if os.path.isfile(representative_source_file):
         # Remove symbolic links before the copy
         try:
-            os.unlink(os.path.join(os.path.expanduser(source_dir), 'fsaverage'))
-            os.unlink(os.path.join(os.path.expanduser(source_dir), 'lh.EC_average'))
-            os.unlink(os.path.join(os.path.expanduser(source_dir), 'rh.EC_average'))
+            os.unlink(os.path.join(os.path.expanduser(source_dir), image_id, 'fsaverage'))
+            os.unlink(os.path.join(os.path.expanduser(source_dir), image_id, 'lh.EC_average'))
+            os.unlink(os.path.join(os.path.expanduser(source_dir), image_id, 'rh.EC_average'))
             for session_id in list_session_ids:
-                os.unlink(os.path.join(os.path.expanduser(source_dir), participant_id + '_' + session_id))
+                os.unlink(os.path.join(os.path.expanduser(source_dir), image_id, participant_id + '_' + session_id))
         except FileNotFoundError as e:
             if e.errno != errno.ENOENT:
                 raise e
 
-        if os.path.isfile(representative_destination_file):
-            if overwrite_caps:
-                shutil.rmtree(destination_dir)
+        if os.path.isfile(representative_destination_file) and overwrite_caps:
+            shutil.rmtree(destination_dir)
             shutil.copytree(os.path.join(source_dir, image_id), destination_dir, symlinks=True)
             print_end_image(image_id)
+
     else:
         now = datetime.datetime.now().strftime('%H:%M:%S')
         cprint('%s[%s] %s does not contain mri/aseg+aparc.mgz file. '
